refactor: make a languages folder and move TypeScripts logo into there
