+++ src/app/logos/frontend/NextJSLogo.tsx
@@ -1,0 +1,15 @@
+import Image from "next/image";
+
+import React from 'react'
+
+export default function NextJSLogo() {
+  return (
+    <Image 
+      src="/images/nextjs_white_icon.png" 
+      width={50}
+      height={50}
+      alt="Next JS logo"
+      style={{ margin: '0 10px', marginBottom: '10px'  }} 
+    />
+  )
+}
